Reorganize grid layout components
