+++ projects/safe/src/lib/components/widgets/grid-layout/add-layout/add-layout.component.ts
@@ -9,7 +9,7 @@ import { SafeGridLayoutService } from '../../../../services/grid-layout.service'
 import { Form } from '../../../../models/form.model';
 import { Layout } from '../../../../models/layout.model';
 import { Resource } from '../../../../models/resource.model';
-import { SafeLayoutModalComponent } from '../../../layout-modal/layout-modal.component';
+import { SafeLayoutModalComponent } from '../layout-modal/layout-modal.component';
 
 /**
  * Data needed for the dialog, should contain a layouts array, a form and a resource
