+++ projects/safe/src/lib/components/widgets/grid-layout/layout-modal/layout-modal.component.html
