+++ projects/safe/src/lib/components/widgets/grid-layout/layouts-parameters/layouts-parameters.component.scss
