+++ projects/safe/src/lib/components/widgets/grid-settings/grid-settings.module.ts
@@ -17,7 +17,7 @@ import { MatTabsModule } from '@angular/material/tabs';
 import { MatChipsModule } from '@angular/material/chips';
 import { SafeButtonModule } from '../../ui/button/button.module';
 import { TranslateModule } from '@ngx-translate/core';
-import { LayoutsParametersModule } from './layouts-parameters/layouts-parameters.module';
+import { LayoutsParametersModule } from '../grid-layout/layouts-parameters/layouts-parameters.module';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { SafeIconModule } from '../../ui/icon/icon.module';
 import { EditorModule, TINYMCE_SCRIPT_SRC } from '@tinymce/tinymce-angular';
