+++ projects/safe/src/lib/safe.module.ts
@@ -32,7 +32,7 @@ import { SafeApplicationsSummaryModule } from './components/applications-summary
 import { SafeSearchMenuModule } from './components/search-menu/search-menu.module';
 import { SafeAggregationBuilderModule } from './components/ui/aggregation-builder/aggregation-builder.module';
 import { SafeTagboxModule } from './components/ui/tagbox/tagbox.module';
-import { SafeLayoutModalModule } from './components/layout-modal/layout-modal.module';
+import { SafeLayoutModalModule } from './components/widgets/grid-layout/layout-modal/layout-modal.module';
 import { SafeSnackbarSpinnerModule } from './components/snackbar-spinner/snackbar-spinner.module';
 import { SafeSkeletonTableModule } from './components/skeleton/skeleton-table/skeleton-table.module';
 import { SafeSkeletonModule } from './directives/skeleton/skeleton.module';
